requirements.txt
colorlog==6.7.0
pip>=21.0,<23.2
pymodbus>=3.1.1
ruff==0.0.263
pytest-homeassistant-custom-component==0.13.23
pytest-asyncio==0.21.0
